fix(ci): ruff format Python + split hook file for Fast Refresh
Reformat server.py and tests/test_server.py with ruff. Split
useFloatingPanel.tsx into three files so every file exports only
components or only non-component code (required by
react-refresh/only-export-components):

- useFloatingPanel.ts — the hook (renamed from .tsx; no JSX)
- ResizeGrip.tsx — the corner-grip component
- floatingPanelGeometry.ts — pure helpers (readPanelGeometry,
  clampGeometryToSafe, rectsOverlap) + shared types

Also drop the unused `h` parameter from the private clampPosition
helper.
