+++ gui/src/hooks/useFloatingPanel.ts
@@ -1,13 +1,9 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-
-// Persisted geometry for a floating panel (drag-to-move + corner resize).
-// Stored under `piperdraw.panel.<id>` in localStorage.
-export interface PanelGeometry {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
+import {
+  type PanelGeometry,
+  panelStorageKey,
+  readPanelGeometry,
+} from "./floatingPanelGeometry";
 
 export interface FloatingPanelOptions {
   id: string;
@@ -16,110 +12,9 @@ export interface FloatingPanelOptions {
   minHeight?: number;
 }
 
-const STORAGE_PREFIX = "piperdraw.panel.";
 const KEEP_ON_SCREEN = 40; // px of the panel that must stay inside the viewport
 
-function storageKey(id: string): string {
-  return STORAGE_PREFIX + id;
-}
-
-function loadGeometry(id: string, fallback: PanelGeometry): PanelGeometry {
-  if (typeof window === "undefined") return fallback;
-  const raw = window.localStorage.getItem(storageKey(id));
-  if (!raw) return fallback;
-  try {
-    const parsed = JSON.parse(raw) as Partial<PanelGeometry>;
-    return {
-      x: Number.isFinite(parsed.x) ? Number(parsed.x) : fallback.x,
-      y: Number.isFinite(parsed.y) ? Number(parsed.y) : fallback.y,
-      width: Number.isFinite(parsed.width) ? Number(parsed.width) : fallback.width,
-      height: Number.isFinite(parsed.height) ? Number(parsed.height) : fallback.height,
-    };
-  } catch {
-    return fallback;
-  }
-}
-
-/**
- * Read a panel's last-persisted geometry without mounting the hook.
- * Returns null if nothing has been persisted yet.
- */
-export function readPanelGeometry(id: string): PanelGeometry | null {
-  if (typeof window === "undefined") return null;
-  const raw = window.localStorage.getItem(storageKey(id));
-  if (!raw) return null;
-  try {
-    const parsed = JSON.parse(raw) as Partial<PanelGeometry>;
-    if (
-      !Number.isFinite(parsed.x) ||
-      !Number.isFinite(parsed.y) ||
-      !Number.isFinite(parsed.width) ||
-      !Number.isFinite(parsed.height)
-    ) {
-      return null;
-    }
-    return {
-      x: Number(parsed.x),
-      y: Number(parsed.y),
-      width: Number(parsed.width),
-      height: Number(parsed.height),
-    };
-  } catch {
-    return null;
-  }
-}
-
-export function rectsOverlap(a: PanelGeometry, b: PanelGeometry): boolean {
-  return (
-    a.x < b.x + b.width &&
-    a.x + a.width > b.x &&
-    a.y < b.y + b.height &&
-    a.y + a.height > b.y
-  );
-}
-
-/**
- * Inset (in px) from each window edge that the panel must avoid — typically
- * picked to clear other fixed UI (top toolbar, bottom hint bar / orientation
- * gizmo / help button).
- */
-export interface SafeMargins {
-  top: number;
-  right: number;
-  bottom: number;
-  left: number;
-}
-
-/**
- * Clamp a panel rectangle to fit inside the window after subtracting `safe`
- * insets. Width/height shrink first (down to `minWidth`/`minHeight`), then x/y
- * shift to keep the rect inside.
- */
-export function clampGeometryToSafe(
-  g: PanelGeometry,
-  safe: SafeMargins,
-  minWidth: number,
-  minHeight: number,
-): PanelGeometry {
-  if (typeof window === "undefined") return g;
-  const minX = safe.left;
-  const minY = safe.top;
-  const maxRight = window.innerWidth - safe.right;
-  const maxBottom = window.innerHeight - safe.bottom;
-  const usableWidth = Math.max(minWidth, maxRight - minX);
-  const usableHeight = Math.max(minHeight, maxBottom - minY);
-  const width = Math.max(minWidth, Math.min(g.width, usableWidth));
-  const height = Math.max(minHeight, Math.min(g.height, usableHeight));
-  let x = g.x;
-  let y = g.y;
-  if (x < minX) x = minX;
-  if (x + width > maxRight) x = maxRight - width;
-  if (y < minY) y = minY;
-  if (y + height > maxBottom) y = maxBottom - height;
-  return { x, y, width, height };
-}
-
-function clampPosition(x: number, y: number, w: number, h: number): { x: number; y: number } {
+function clampPosition(x: number, y: number, w: number): { x: number; y: number } {
   if (typeof window === "undefined") return { x, y };
   const maxX = window.innerWidth - KEEP_ON_SCREEN;
   const maxY = window.innerHeight - KEEP_ON_SCREEN;
@@ -141,21 +36,21 @@ function clampPosition(x: number, y: number, w: number, h: number): { x: number;
 export function useFloatingPanel(opts: FloatingPanelOptions) {
   const { id, defaultGeometry, minWidth = 220, minHeight = 160 } = opts;
 
-  const [geom, setGeom] = useState<PanelGeometry>(() =>
-    loadGeometry(id, defaultGeometry),
+  const [geom, setGeom] = useState<PanelGeometry>(
+    () => readPanelGeometry(id) ?? defaultGeometry,
   );
 
   // Persist on every change (cheap; geometry is 4 numbers).
   useEffect(() => {
     if (typeof window === "undefined") return;
-    window.localStorage.setItem(storageKey(id), JSON.stringify(geom));
+    window.localStorage.setItem(panelStorageKey(id), JSON.stringify(geom));
   }, [id, geom]);
 
   // Keep the panel reachable if the window shrinks below its position.
   useEffect(() => {
     const onResize = () => {
       setGeom((g) => {
-        const c = clampPosition(g.x, g.y, g.width, g.height);
+        const c = clampPosition(g.x, g.y, g.width);
         return c.x === g.x && c.y === g.y ? g : { ...g, ...c };
       });
     };
@@ -191,7 +86,6 @@ export function useFloatingPanel(opts: FloatingPanelOptions) {
         d.origX + (e.clientX - d.startX),
         d.origY + (e.clientY - d.startY),
         g.width,
-        g.height,
       );
       return { ...g, ...next };
     });
@@ -276,7 +170,7 @@ export function useFloatingPanel(opts: FloatingPanelOptions) {
       setGeom((g) => {
         const width = Math.max(minWidth, patch.width ?? g.width);
         const height = Math.max(minHeight, patch.height ?? g.height);
-        const pos = clampPosition(patch.x ?? g.x, patch.y ?? g.y, width, height);
+        const pos = clampPosition(patch.x ?? g.x, patch.y ?? g.y, width);
         return { x: pos.x, y: pos.y, width, height };
       });
     },
@@ -284,33 +178,4 @@ export function useFloatingPanel(opts: FloatingPanelOptions) {
   );
 
   return { containerStyle, dragHandleProps, resizeGripProps, geometry: geom, setGeometry };
-}
-
-export function ResizeGrip(props: {
-  onPointerDown: (e: React.PointerEvent<HTMLDivElement>) => void;
-  onPointerMove: (e: React.PointerEvent<HTMLDivElement>) => void;
-  onPointerUp: (e: React.PointerEvent<HTMLDivElement>) => void;
-  onPointerCancel: (e: React.PointerEvent<HTMLDivElement>) => void;
-}) {
-  return (
-    <div
-      role="separator"
-      aria-label="Resize panel"
-      title="Drag to resize"
-      {...props}
-      style={{
-        position: "absolute",
-        right: 0,
-        bottom: 0,
-        width: 14,
-        height: 14,
-        cursor: "nwse-resize",
-        touchAction: "none",
-        backgroundImage:
-          "linear-gradient(135deg, transparent 0 6px, #999 6px 7px, transparent 7px 10px, #999 10px 11px, transparent 11px)",
-        borderBottomRightRadius: 8,
-        zIndex: 1,
-      }}
-    />
-  );
 }
